moved liquid stake test
